249: fixed module interface for inject epoch hooks and some fixes for generation committee (#251)
* 249: fixed module interface for di inject hook and some fixes for generation committee

* chore(review): fixed consistent interface naming

* chore(review): fixed type consistency for epoch

+++ x/committee/types/repository.go
@@ -3,12 +3,12 @@ package types
 import sdk "github.com/cosmos/cosmos-sdk/types"
 
 type Repository interface {
-	GetLastEpoch(ctx sdk.Context) (uint32, error)
-	SetLastEpoch(ctx sdk.Context, epoch uint32) error
+	GetLastEpoch(ctx sdk.Context) (int64, error)
+	SetLastEpoch(ctx sdk.Context, epoch int64) error
 
-	GetCommittee(ctx sdk.Context, epoch uint32) (Committee, error)
+	GetCommittee(ctx sdk.Context, epoch int64) (Committee, error)
 	GetLastCommittee(ctx sdk.Context) (Committee, error)
-	SetCommittee(ctx sdk.Context, epoch uint32, committee Committee) error
-	HasCommittee(ctx sdk.Context, epoch uint32) (bool, error)
+	SetCommittee(ctx sdk.Context, epoch int64, committee Committee) error
+	HasCommittee(ctx sdk.Context, epoch int64) (bool, error)
 	IterateCommittees(ctx sdk.Context, fn func(committee Committee) error) error
 }
